+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Misc
 
+* Add command line argument `--seed-count n`
+  * This allows generating `n` seeds at once, and is intended as a mass-testing feature. Despite this, it is still available to the user.
 * Add EDC / ECC (error detection code, error correction code) generation
   * This should allow more emulators, most importantly bizhawk, to properly play ROMs of this randomizer